+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # 20-10-2024
 
 * Added a board representation in a form of bitboard.
+
+# 28-10-2024
+
+* Extracted the pieces representations to a union so they can be accessed by name and by offset.
+* Implemented a Move struct that changes the board state.